refactor(models): extract provider_ai_model_display and thin base conversation view
